refactor: modularize summary methods and math utils
