+++ js/game-data/detail-stat-list.js
@@ -1,0 +1,47 @@
+const DetailStatList = [
+  'Takedowns',
+  'SoloKill',
+  'Assists',
+  'Deaths',
+  'KillParticipation', // special case
+  'KDA',
+  'HighestKillStreak',
+  'VengeancesPerformed',
+  'TimeSpentDead',
+  'OutnumberedDeaths',
+  'EscapesPerformed',
+  'TeamfightEscapesPerformed',
+  'HeroDamage',
+  'DPM',
+  'damageDonePerDeath',
+  'TeamfightHeroDamage',
+  'SiegeDamage',
+  'StructureDamage',
+  'MinionDamage',
+  'SummonDamage',
+  'CreepDamage',
+  'Healing',
+  'HPM',
+  'healingDonePerDeath',
+  'TeamfightHealingDone',
+  'SelfHealing',
+  'ProtectionGivenToAllies',
+  'ClutchHealsPerformed',
+  'DamageTaken',
+  'damageTakenPerDeath',
+  'TeamfightDamageTaken',
+  'TimeCCdEnemyHeroes',
+  'TimeRootingEnemyHeroes',
+  'TimeSilencingEnemyHeroes',
+  'TimeStunningEnemyHeroes',
+  //'TimeOnPoint',
+  'OnFireTimeOnFire',
+  'ExperienceContribution',
+  'XPM',
+  'MercCampCaptures',
+  //'TownKills',
+  'WatchTowerCaptures'
+  //'Role'
+];
+
+module.exports = DetailStatList;
