+++ js/game-data/map-stats.js
@@ -1,0 +1,18 @@
+const PerMapStatList = {
+  "Towers of Doom" : ["AltarDamageDone"],
+  "Battlefield of Eternity" : ["DamageDoneToImmortal"],
+  "Dragon Shire" : ["DragonNumberOfDragonCaptures", "DragonShrinesCaptured"],
+  "Blackheart's Bay" : ["BlackheartDoubloonsCollected", "BlackheartDoubloonsTurnedIn"],
+  "Haunted Mines" : ["MinesSkullsCollected"],
+  "Infernal Shrines" : ["DamageDoneToShrineMinions"],
+  "Garden of Terror" : ["GardensPlantDamage", "GardensSeedsCollected"],
+  "Tomb of the Spider Queen" : ["GemsTurnedIn"],
+  "Warhead Junction" : ["NukeDamageDone"],
+  "Cursed Hollow" : ["CurseDamageDone"],
+  "Volskaya Foundry" : [],
+  "Sky Temple" : ["TimeInTemple"],
+  "Braxis Holdout" : ["DamageDoneToZerg"],
+  "Hanamura" : []
+};
+
+module.exports = PerMapStatList;
